# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/features/{feature}": {
            "put": {
                "operationId": "features.update",
                "summary": "Update an existing feature",
                "tags": [
                    "Feature"
                ],
                "parameters": [
                    {
                        "name": "feature",
                        "in": "path",
                        "required": true,
                        "description": "The feature i id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreFeatureRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`FeatureResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/FeatureResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "features.destroy",
                "summary": "Soft-delete an existing feature",
                "tags": [
                    "Feature"
                ],
                "parameters": [
                    {
                        "name": "feature",
                        "in": "path",
                        "required": true,
                        "description": "The feature i id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
        "/features/{feature}/restore": {
            "post": {
                "operationId": "features.restore",
                "summary": "Restore a soft-deleted feature",
                "tags": [
                    "Feature"
                ],
                "parameters": [
                    {
                        "name": "feature",
                        "in": "path",
                        "required": true,
                        "description": "The feature i id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`FeatureResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/FeatureResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FeatureResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "alias": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "parent": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "route": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "icon": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "order": {
                        "type": "integer"
                    },
                    "show_on_sidebar": {
                        "type": "boolean"
                    },
                    "created_by": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_by": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "deleted_by": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "deleted_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "name",
                    "alias",
                    "type",
                    "parent",
                    "description",
                    "route",
                    "icon",
                    "order",
                    "show_on_sidebar",
                    "created_by",
                    "created_at",
                    "updated_by",
                    "updated_at",
                    "deleted_by",
                    "deleted_at"
                ],
                "title": "FeatureResource"
            },
...
            "StoreFeatureRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "alias": {
                        "type": "string",
                        "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                        "maxLength": 150
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "menu",
                            "crud",
                            "filter"
                        ]
                    },
                    "parent": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 150
                    },
                    "route": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 250
                    },
                    "icon": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 50
                    },
                    "order": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "minimum": 1,
                        "maximum": 32767
                    },
                    "show_on_sidebar": {
                        "type": "boolean"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 100
                    }
                },
                "required": [
                    "name",
                    "alias",
                    "type"
                ],
                "title": "StoreFeatureRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}